QA hardening + architecture doc + "how to open a .glb"
QA pass over the decompose feature (11 commits this cycle):
- decompose_export_turnaround: real path-component sanitisation (reject
  / \ : .. and non-[A-Za-z0-9_-], len cap), 400-frame / 40MB-per-frame
  caps against a runaway IPC payload.
- runTurnaround: per-object try/catch so one bad GLB doesn't abort the
  whole export; slug clamped to 40 chars with a fallback.
- fal_edit_image: drop the dead `let _ = rest`.
- clippy: CatalogCache type alias; drop a redundant u32 cast in
  decompose_setup. Zero clippy warnings in the decompose files now.
- Verified: cargo + tsc + clippy clean; Python stub run OK; a real
  2-object Tripo run + turnaround export produced a valid 4.7MB zip
  (10x 1024^2 PNG + manifest).

Docs:
- docs/decompose-architecture.md — full developer reference: data flow,
  every command, on-disk state, the Python contract, the fan-out, the
  scene.json schema, frontend notes, known limits, extension points.
- In-app "How to use these files" + tutorial now answer "how do I open a
  .glb" — in-app ▶, Windows 3D Viewer / Paint 3D, browser viewers, VS
  Code glTF Tools, Blender, ModelForge. Tutorial documents Turnaround
  (.zip).

+++ src-tauri/src/decompose_setup.rs
@@ -211,7 +211,7 @@ async fn run_streamed(
     let mut seen: u32 = 0;
     while let Ok(Some(line)) = lines.next_line().await {
         seen += 1;
-        let frac = (seen.min(40) as u32 * span as u32) / 40;
+        let frac = seen.min(40) * span as u32 / 40;
         let pct = (base as u32 + frac).min(99) as u8;
         let short: String = line.trim().chars().take(90).collect();
         if !short.is_empty() {
